Move keybindings and key to core/src/browser
Since these rely on DOM events.

Fixes: #434
