+++ packages/core/src/browser/saveable.ts
@@ -7,7 +7,8 @@
 
 import { Widget } from '@phosphor/widgets';
 import { Message } from '@phosphor/messaging';
-import { Event, MaybePromise, Key } from '../common';
+import { Event, MaybePromise } from '../common';
+import { Key } from './keys';
 import { AbstractDialog } from './dialogs';
 
 export interface Saveable {
