+++ packages/core/src/browser/tree/tree-widget.ts
@@ -9,7 +9,8 @@ import { injectable, inject } from "inversify";
 import { Message } from "@phosphor/messaging";
 import { ElementExt } from "@phosphor/domutils";
 import { h, ElementAttrs, ElementInlineStyle } from "@phosphor/virtualdom";
-import { Disposable, Key, MenuPath } from "../../common";
+import { Disposable, MenuPath } from "../../common";
+import { Key } from "../keys";
 import { ContextMenuRenderer } from "../context-menu-renderer";
 import { StatefulWidget } from '../shell';
 import { VirtualWidget, VirtualRenderer, SELECTED_CLASS, COLLAPSED_CLASS } from "../widgets";
