+++ packages/cpp/src/browser/cpp-keybinding.ts
@@ -9,7 +9,7 @@ import { injectable, inject } from "inversify";
 import { EditorManager } from "@theia/editor/lib/browser";
 import {
     KeybindingContext, Keybinding, KeybindingContribution, KeybindingRegistry
-} from "@theia/core/lib/common";
+} from "@theia/core/lib/browser";
 import { editorContainsCppFiles } from "./cpp-commands";
 
 @injectable()
