+++ packages/file-search/src/browser/file-search-frontend-module.ts
@@ -6,8 +6,8 @@
  */
 
 import { ContainerModule, interfaces } from 'inversify';
-import { CommandContribution, KeybindingContribution } from "@theia/core/lib/common";
-import { WebSocketConnectionProvider } from '@theia/core/lib/browser';
+import { CommandContribution } from "@theia/core/lib/common";
+import { WebSocketConnectionProvider, KeybindingContribution } from '@theia/core/lib/browser';
 import { QuickFileOpenFrontendContribution } from './quick-file-open-contribution';
 import { QuickFileOpenService } from './quick-file-open';
 import { fileSearchServicePath, FileSearchService } from '../common/file-search-service';
