+++ packages/markers/src/browser/problem/problem-frontend-module.ts
@@ -9,8 +9,8 @@ import { ContainerModule } from 'inversify';
 import { ProblemWidget } from './problem-widget';
 import { ProblemContribution } from './problem-contribution';
 import { createProblemWidget } from './problem-container';
-import { CommandContribution, MenuContribution, KeybindingContribution } from "@theia/core/lib/common";
-import { FrontendApplicationContribution } from '@theia/core/lib/browser';
+import { CommandContribution, MenuContribution } from "@theia/core/lib/common";
+import { FrontendApplicationContribution, KeybindingContribution } from '@theia/core/lib/browser';
 import { ProblemManager } from './problem-manager';
 import { PROBLEM_KIND } from '../../common/problem-marker';
 import { WidgetFactory } from '@theia/core/lib/browser/widget-manager';
